requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==24.3
django-grappelli==4.0.1
django-js-asset==2.2.0
django-nine==0.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
elastic-transport==8.15.1
elasticsearch==8.15.1
elasticsearch-dsl==8.15.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
